Moved the login in the `MainWindow.xaml.cs` into a controller structure
